+++ src/app/components/accueil/bienvenu/bienvenu.component.html
@@ -1,3 +1,4 @@
+
 <div class="container">
   <div class="leftSide">
     <mat-card>
@@ -22,6 +23,7 @@
         <div class="texte">
           <h4>Recompenses</h4>
           <p>Prix de l'innovation 2019. Nous en sommes très fiers</p>
+        <p></p>
         </div>
 
         <div>
@@ -32,15 +34,17 @@
       </div>
     </mat-tab>
 
-    <!-- ******* -->
-
     <mat-tab label="Les dernières actus">
       <div class="section mat-elevation-z4">
         <div class="texte">
           <h4>Démarrage</h4>
-      <p>dévut de l'expérimentation avec nos 35 familles.</p>
+          <p>Début de l'expérimentation avec nos 35 familles.</p>
         </div>
-    <div><img src="http://via.placeholder.com/60x60" alt="placeholder"></div>
+        <div>
+          <img src="http://via.placeholder.com/60x60" alt="placeholder">
+        </div>
+            <!--  compte à rebours debut-->
+
       </div>
     </mat-tab>
 
@@ -50,7 +54,7 @@
           <h4>{{item.username}} </h4>
           <p>{{item.description}}</p>
         </div>
-    <div>
+        <div>
           <img src="http://via.placeholder.com/60x60" alt="placeholder">
         </div>
       </div>
