+++ src/app/components/accueil/bienvenu/bienvenu.component.ts
@@ -2,14 +2,15 @@ import { Witness } from './../../../shared/models/witness';
 import { WitnessService } from './../../../shared/services/witness.service';
 import { Component, OnInit } from '@angular/core';
 
-
 @Component({
   selector: 'app-bienvenu',
   templateUrl: './bienvenu.component.html',
   styleUrls: ['./bienvenu.component.scss']
 })
+
+
 export class BienvenuComponent implements OnInit {
- witness: Witness[];
+  witness: Witness[];
   constructor(private witnessService: WitnessService) { }
 
   ngOnInit() {
@@ -18,4 +19,3 @@ export class BienvenuComponent implements OnInit {
     });
   }
 }
-
