+++ src/app/shared/services/clock.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ClockService } from './clock.service';
+
+describe('ClockService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ClockService = TestBed.get(ClockService);
+    expect(service).toBeTruthy();
+  });
+});
