+++ src/index.html
@@ -10,13 +10,22 @@
   <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css?family=Sarabun:300,600&display=swap" rel="stylesheet">
-  <!-- dans le cas de l'utilisation autre que de l'iframe (intégration de droite) -->
-  <div id="fb-root"></div>
-  <script async defer crossorigin="anonymous" src="https://connect.facebook.net/fr_FR/sdk.js#xfbml=1&version=v5.0"></script>
+
+  <!-- head horloge 2 debut -->
+  <!-- <link rel="stylesheet" media="screen" type="text/css" title="Style" href="style/styles.css" /> -->
+  <!-- <script type="text/javascript" src="src/app/components/accueil/bienvenu/index2.js"></script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.8.3/jquery.min.js"></script>
+  <script src="http://code.jquery.com/ui/1.9.2/jquery-ui.js"></script> -->
+  <!-- head horloge 2 fin -->
 
 
 </head>
 <body>
+
+  <!-- dans le cas de l'utilisation autre que de l'iframe (intégration de la deuxième fb) -->
+  <div id="fb-root"></div>
+  <script async defer crossorigin="anonymous" src="https://connect.facebook.net/fr_FR/sdk.js#xfbml=1&version=v5.0"></script>
+
   <app-root></app-root>
 </body>
 </html>
