+++ src/main.ts
@@ -2,7 +2,7 @@ import 'hammerjs';
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { AppModule } from './app/shared/services/app.module';
 import { environment } from './environments/environment';
 
 if (environment.production) {
